# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"MessageStatus": {
				"description": "0: Sent, 1: Received, 2: Seen, 3: Deleted",
				"enum": [
					0,
					1,
					2,
					3
				],
				"type": "number"
			},
			"Message": {
				"properties": {
					"id": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/MessageStatus"
					}
				},
				"required": [
					"id",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Message.message_": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NewMessage": {
				"$ref": "#/components/schemas/Pick_Message.message_"
			}
		},
		"securitySchemes": {
			"AuthToken": {
				"type": "apiKey",
				"name": "x-auth",
				"in": "header",
				"description": "User Auth Token"
			}
		}
	},
	"info": {
		"title": "Node Typescript Boilerplate",
		"version": "1.0.0",
		"description": "Default Node Typescript Boilerplate description",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LizardKing"
		}
	},
	"paths": {
		"/messages/{messageId}": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					}
				},
				"description": "Get a test message.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/messages": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"description": "Send a test message. \\n\nExample AuthToken: abc123456",
				"security": [
					{
						"authToken": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewMessage"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}